secure email to use email instead of username

+++ logic/account/account.py
@@ -1,3 +1,5 @@
+import os
+
 from .__init__ import *
 
 
@@ -11,6 +13,9 @@ class RoomAccount:
     * authenticate - Confirm if user account exists -> false | true.
     * deactivate - Pend user-account for deactivation after number of days.
     """
+
+    # application root directory
+    root_dir = os.getcwd()
 
     def __init__(self, profile: dict):
         self.profile: dict = profile  # user's full profile
@@ -63,31 +68,72 @@ class RoomAccount:
     async def authenticate(self):
         file = f'{self.account_directory}/{self.username}.db'
 
-        if os.path.exists(file):
-            database = sqlite3.connect(file)
-            cursor = database.cursor()
+        if self.username != '':
+            if os.path.exists(file):
+                database = sqlite3.connect(file)
+                cursor = database.cursor()
 
-            # passwords
-            input_password = self.password.encode('utf-8')
-            local_password: str = cursor.execute("""
-            select json_extract(data, '$.password') from profile;
-            """).fetchone()[0].encode('utf-8')
+                # passwords
+                input_password = self.password.encode('utf-8')
+                local_password: str = cursor.execute("""
+                select json_extract(data, '$.password') from profile;
+                """).fetchone()[0].encode('utf-8')
 
-            # compare hashed passwords
-            try:
-                if bcrypt.checkpw(input_password, local_password):
-                    return account_access_granted
-                else:
-                    return account_access_denied_password
-            except ValueError as error:
-                return account_access_denied_passwordhash
+                # compare hashed passwords
+                try:
+                    if bcrypt.checkpw(input_password, local_password):
+                        return {"authentication_result": account_access_granted}
+                    else:
+                        return {"authentication_result": account_access_denied_password}
+                except ValueError as error:
+                    return {"authentication_result": account_access_denied_passwordhash}
 
-        else:
-            if os.path.exists(self.account_directory):
-                shutil.rmtree(self.account_directory)
-                return account_dir_exists_restored
             else:
-                return account_exists_false
+                if os.path.exists(self.account_directory):
+                    shutil.rmtree(self.account_directory)
+                    return {"authentication_result": account_dir_exists_restored}
+                else:
+                    return {"authentication_result": account_exists_false}
+        else:
+            if self.email != '':
+                try:
+                    accounts_parent_dir = os.listdir(f"system/user/account")
+                    for directory in accounts_parent_dir:
+                        if os.path.isdir(f'system/user/account/{directory}'):
+                            os.chdir(f'system/user/account/{directory}')
+                            if os.path.isfile(f'{directory}.db'):
+                                database = sqlite3.connect(f'{directory}.db')
+                                cursor = database.cursor()
+
+                                # user profile with minimum data
+                                user_profile: dict = json.loads(cursor.execute("""
+                                select json_extract(data, '$') from profile;
+                                """).fetchone()[0])
+
+                                # restore root-dir as work-dir
+                                os.chdir(self.root_dir)
+
+                                # passwords
+                                input_password = self.password.encode('utf-8')
+                                local_password: str = user_profile['password'].encode('utf-8')
+
+                                TODO: "if no account have a matching email then return account_exists_false_value"
+
+                                # check for account with matching email & password
+                                if user_profile['email']:
+                                    if user_profile['email'] == self.email:
+                                        if bcrypt.checkpw(input_password, local_password):
+                                            return {
+                                                "username": user_profile['username'],
+                                                "authentication_result": account_access_granted}
+                                        else:
+                                            return {"authentication_result": account_access_denied_password}
+                except FileNotFoundError:
+                    # account-dir does not exist
+                    return {"authentication_result": account_exists_false}
+            else:
+                # username & email have no values
+                return {"authentication_result": account_exists_false}
 
     # pend account for deactivation
     async def deactivate(self):
